A simple homework of socket
There is huge amount of bugs , i even found some, and it does work. So Plz not use it in any program. only help you to learn it. The socket part is part copy from a website and i marked reference. If it will Infringe on your copyright ，plz contact me.